requirements.txt
anyio @ file:///croot/anyio_1706220167567/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///work/ci_py311/argon2-cffi-bindings_1676823553406/work
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru @ file:///croot/async-lru_1699554519285/work
attrs @ file:///croot/attrs_1695717823297/work
Babel @ file:///work/ci_py311/babel_1676825020543/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
cobra_db==0.4.6
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.0
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
deid==0.2.37
dnspython==2.6.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///work/ci_py311_2/python-fastjsonschema_1679340124475/work
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
gdcm==1.1
h11 @ file:///croot/h11_1706652277403/work
httpcore @ file:///croot/httpcore_1706728464539/work
httpx @ file:///croot/httpx_1723474802858/work
huggingface-hub==0.25.1
idna @ file:///croot/idna_1714398848350/work
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1726064227985/work
ipywidgets @ file:///croot/ipywidgets_1709574692113/work
jedi @ file:///croot/jedi_1721058342488/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///croot/jsonschema_1699041609003/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter @ file:///croot/jupyter_1707947101020/work
jupyter-console @ file:///croot/jupyter_console_1679999630278/work
jupyter-events @ file:///croot/jupyter_events_1718738097486/work
jupyter-lsp @ file:///croot/jupyter-lsp-meta_1699978238815/work
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server @ file:///croot/jupyter_server_1718827083372/work
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1686870725608/work
jupyterlab @ file:///croot/jupyterlab_1725895214311/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1709322880313/work
jupyterlab_server @ file:///croot/jupyterlab_server_1725865349919/work
kiwisolver==1.4.7
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.9.2
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mistune @ file:///work/ci_py311/mistune_1676823683049/work
mpmath==1.3.0
nbclient @ file:///croot/nbclient_1698934205032/work
nbconvert @ file:///croot/nbconvert_1699022732553/work
nbformat @ file:///croot/nbformat_1694616755618/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.3
notebook @ file:///croot/notebook_1725954770513/work
notebook_shim @ file:///croot/notebook-shim_1699455894279/work
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
overrides @ file:///croot/overrides_1699371140756/work
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.2.3
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==10.4.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
ply==3.11
prometheus-client @ file:///work/ci_py311_2/prometheus_client_1679340520168/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
psutil @ file:///work/ci_py311_2/psutil_1679337388738/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyaml-env==1.2.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.20.0
pydicom==2.4.4
Pygments @ file:///croot/pygments_1684279966437/work
pylibjpeg==2.0.1
pylibjpeg-libjpeg==2.2.0
pymongo==4.9.1
pyparsing==3.1.4
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger @ file:///croot/python-json-logger_1683823803357/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1698096049011/work
pyzmq @ file:///croot/pyzmq_1705605076900/work
qtconsole @ file:///croot/qtconsole_1726836502164/work
QtPy @ file:///croot/qtpy_1700144840038/work
referencing @ file:///croot/referencing_1699012038513/work
requests @ file:///croot/requests_1721410876868/work
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
rpds-py @ file:///croot/rpds-py_1698945930462/work
safetensors==0.4.5
Send2Trash @ file:///croot/send2trash_1699371139552/work
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///croot/sniffio_1705431295498/work
soupsieve @ file:///croot/soupsieve_1696347547217/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.13.3
terminado @ file:///work/ci_py311/terminado_1677696151350/work
timm==1.0.9
tinycss2 @ file:///work/ci_py311/tinycss2_1676823757641/work
torch==2.4.1
torchvision==0.19.1
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.66.5
traitlets @ file:///croot/traitlets_1718227057033/work
triton==3.0.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.2
urllib3 @ file:///croot/urllib3_1718912636303/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///croot/websocket-client_1715878298792/work
widgetsnbextension @ file:///croot/widgetsnbextension_1709322880396/work
